pyproject.toml
[tool.poetry]
name = "image-gallery"
version = "0.1.0"
description = ""
authors = ["Eugene <giveuper39@gmail.com>"]
packages = [{include = "image_gallery", from = "."}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.2"
boto3 = "^1.40.74"
python-dotenv = "^1.2.1"
pillow = "^12.0.0"
gunicorn = "^23.0.0"
flask-httpauth = "^4.8.0"


[tool.poetry.group.dev.dependencies]
black = "^25.11.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 89

[tool.poetry.scripts]
startapp = "image_gallery.main:create_app"
